Edit package and class names
Package name: seedu.duke

Package name should follow name of program instead to prevent confusion
as the program scales up.

Let's,
* change package name to seedu.meal360
* change class name to Meal360
* edit application and shadowJar of build.gradle
