+++ src/test/java/seedu/meal360/Meal360Test.java
@@ -1,10 +1,10 @@
-package seedu.duke;
+package seedu.meal360;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
-class DukeTest {
+class Meal360Test {
     @Test
     public void sampleTest() {
         assertTrue(true);
