+++ text-ui-test/EXPECTED.TXT
@@ -1,9 +1,8 @@
 Hello from
- ____        _        
-|  _ \ _   _| | _____ 
-| | | | | | | |/ / _ \
-| |_| | |_| |   <  __/
-|____/ \__,_|_|\_\___|
+ __  __          _ ____  __  __
+|  \/  |___ __ _| |__ / / / /  \
+| |\/| / -_) _` | ||_ \/ _ \ () |
+|_|  |_\___\__,_|_|___/\___/\__/
 
 What is your name?
 Hello James Gosling
